Script develop to populate the Users table from the backend
