+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "cross-env DB_ENV=testing jest --watch --verbose",
     "server": "nodemon index.js",
-    "start": "node index.js"
+    "start": "node index.js",
+    "cleanData": "cd routes/ && node dataCleaner.js"
   },
   "repository": {
     "type": "git",
