+++ seeds/001_users_seed.js
@@ -1,3 +1,5 @@
+
+
 exports.seed = function(knex, Promise) {
   // Deletes ALL existing entries
   return knex("users")
